Create Common library and move ABIArgInfo to it
